Move external contract call tests to a seperate file
